Remove unused by_geozone_id scope from Poll
The "by_geozone_id" scope in Poll is no longer used anywhere in the code.
It was first introduced in commit 20cb0440150e ("adds search and filter
for poll questions") and later moved to the Poll model in commit
d02450596048 ("moves geozones from poll question to poll in models"),
but has never been referenced since.