Use the default view mode icon by default
In the `default_view_mode?` helper we assume that the view mode is the
default one unless the minimal mode is explicitly specified. So now
we're applying the same logic to the CSS. This is usually how the
cascade works: you define a default style, and then overwrite it with a
specific one.

This means we no longer need to specify the `default` HTML class when
the view mode is the default one.