Use aria-current instead of is-active
This way people using screen readers also have access to this
information.

To simplify the logic, we're using the `link_list` helper. This changes
the existing behaviour; previously, the current element wasn't a link.
We're making it a link because that's what we usually do.

Besides, the previous behaviour was a bit inaccessible. The item that
was a link had the color we usually use on regular text, while the item
that wasn't a link had the color we usually use on links. And the icons
weren't part of the link; that means it was harder to hit the link for
people using touchsceens.