Use `normalizes` to clean document numbers
Since we no longer have to wait for validation to normalize this
attribute, we can simplify the tests a little bit.

Note that in Rails 7.1 the `normalizes` method isn't available in
ActiveModel. That means we can't use it in `Officing::Residence` and
`Verification::Residence`. For that, we'll have to wait for Rails 8.1.

Note that there's still some duplication and we could certainly extract
the regular expression to a shared class. We'll do so (unless we forget)
when we upgrade to Rails 8.1 and start using ActiveModel normalizations.