Upgrade to Rails 8.0 TODO check production logs formatter
Note that Rails 8.0 groups the parameters sent in our custom information
texts form differently, so we need to update the `I18nContent.update`
method accordingly.

Compared to the code generated by `rails app:update`, we're reverting a
change in `bin/setup` that used `bin/dev` insted of `bin/rails restart`.
The reason is that we currently don't use `bin/dev` and it fails when a
server is already running.

We're also keeping using `null_store` in development when
`tmp/caching-dev.txt` isn't present; the opposite was a bit confusing in
my opinion.

The `config.assets.quiet` option in the development environment was
removed in Rails 8.0 but added back in Rails 8.1 [1], so we're keeping
it.

Since `config.assume_ssl` is true by default on production, we're
setting it to the same value as `force_ssl`.

We also keep serving static assets with Nginx by default and rotating
the logger on production.

[1] Pull request 53627 in https://github.com/rails/rails/