Upgrade to Rails 8.0
Compared to the code generated by `rails app:update`, we're reverting a
change in `bin/setup` that used `bin/dev` insted of `bin/rails restart`.
The reason is that we currently don't use `bin/dev` and it fails when a
server is already running.

We're also keeping using `null_store` in development when
`tmp/caching-dev.txt` isn't present; the opposite was a bit confusing in
my opinion.

In the test environment, we keep disabling the cache to keep the
existing behavior. Enabling the cache might catch bugs related to cache
issues, but we need to check whether there are side effects.

The `config.assets.quiet` option in the development environment was
removed in Rails 8.0 but added back in Rails 8.1 [1], so we're keeping
it.

Since `config.assume_ssl` is true by default on production, we're
setting it to the same value as `force_ssl`.

We also keep serving static assets with Nginx by default and rotating
the logger on production.

Note we aren't using the `ActiveSupport::TaggedLogging.logger` method.
We didn't manage to do so and still keep using a formatter with the PID
and timestamp. When doing so, we got an error:

```
<NoMethodError: undefined method 'push_tags' for an instance of
Logger::Formatter>
```

As mentioned in commit cb0084c41, we no longer need to keep using
`connection_pool` 2.x, so we're upgrading it.

[1] Pull request 53627 in https://github.com/rails/rails/