Move active storage service config to environments
This is the default behavior. We used to have environment files for
staging and preproduction, so we added it to `config/application.rb`.
However, now that we've got rid of staging and preproduction, having it
in `config/environments/` is more convenient because that's what we get
every time we upgrade to a new version of Rails and run `rails
app:upgrade`.