Make `to_time` preserve the timezone of the receiver
This option was removed in Rails 8.1 [1], so we don't have much of a
choice regarding whether to adopt the new `:zone` value. We were also
getting a warning:

```
DEPRECATION WARNING: `to_time` will always preserve the full timezone
rather than offset of the receiver in Rails 8.1. To opt in to the new
behavior, set `config.active_support.to_time_preserves_timezone =
:zone`.
```

We don't use the `to_time` methods ourselves, so we'll have to make sure
it don't cause any trouble in any of our dependencies.

[1] Commit c79dc282d687 in https://github.com/rails/rails/