Check LineLength rule in production.rb
With the new changes in this file, we don't have lines exceeding this
length anymore, so we can enable the rule.

The `environments/staging.rb` file also follows this rule since commit
b5b03e449, so we're no longer excluding it from this check either.

We're also checking `Layout/MultilineMethodCallIndentation`, which
we've been following for a long time now.