Make the `secrets` method compatible with Rails 7.2
We're about to upgrade to Rails 7.2, and in this version the `secrets`
method has been removed. Since we don't plan to start using credentials
in the foreseeable future, we're using a simplified version of the
`secrets` method found in Rails 7.1.

We also need to overwrite the `secret_key_base` method so it reads the
one in `config/secrets.yml`. However, doing so is tricky since this
method contains a few conditionals related to the current environment
and to certain environment variables. Currently the logic of this method
(extracted from the source code of Rails 7.2) is:

```
if ENV["SECRET_KEY_BASE_DUMMY"]
  generate_local_secret
else
  ENV["SECRET_KEY_BASE"] ||
    Rails.application.credentials.secret_key_base ||
    (Rails.env.local? && generate_local_secret)
end
```

Instead, we're overwriting the `credentials` method so it returns
secrets instead; since the `secret_key_base` method calls
`credentials.secret_key_base` when running the application on
production, overwriting the `credentials` method will cause the
application to get the secret key base from the `config/secrets.yml`
file. Since we don't use credentials at all, this change doesn't affect
the rest of the application.

To avoid confusion, we're changing the Devise initializer so it doesn't
reference the `secrets.secret_key_base` method but uses
`application.secret_key_base` instead.