Use a shared component to render basic author info
We were using the same code in six different places.

The only change we're doing is that the author information shown in the
list of investements was slightly different and now is the same as
everywhere else. We believe it was a typo due to code duplication and
that there was no real intention behind the changes in the official
level and organization logic for investments.

For example, the `official?` condition on investments was added in
commit be6d7a1d3 on June 3, 2016, copied from proposals and debates.
However, by the time it was finally merged in commit 14bb9eef8, the
logic in proposals and debates had changed to
`display_official_position_badge?` (commit 9e910a834).

Not sure why the organization was shown in debates, proposals and
legislation proposals but not in investments, though, but it was
probably another typo.

Note we're adding `Basic` to the component name because there's already
a `shared/_author_info` partial.

Also note that, in featured proposals and proposals summary, we're still
hiding the organization information. Not sure whether this information
should be displayed in these cases; we're doing it to keep the existing
behavior, since the information displayed in those sections is different
than the one displayed in other places.