Use rails defaults for puma and database threads
Back in commit 1e32455cd, we used puma's default values for threads.
This was a mistake, though; for Rails application, the maximum number of
threads should be at most the number defined in the database pool. The
Rails-generated puma configuration says:

> Any libraries that use a connection pool or another resource pool
> should be configured to provide at least as many connections as the
> number of threads. This includes Active Record's `pool` parameter in
> `database.yml`.

So we now set the number of threads to 3 and the size of the database
poll to 5 by default, which is what Rails 7.2 does. Note that
overwriting `RAILS_MAX_THREADS` would result in the database pool having
the same size as the number of threads. In my humble opinion, this is a
bit confusing, and there's currently debate about what Rails should do
in the future [1].

[1] Pull request 55687 in https://github.com/rails/rails