Don't link to the admin area from the moderation area
Not all moderators are necessarily administrators, and we were sending
them to an area they couldn't access. So, just like we're doing in all
the other sections in the moderation area, we're linking to the public
area instead.

There's currently a test that checks what happens when you click that
link, but only for comments. In the next commit we'll adapt it so it
works on debates, proposals and budget investments too.