Remove code duplication in moderation tables
With a couple of small differences, we were using the same code in five
different places.

Note we're renaming the `field_name` method to `ids_field_name` because
it's more readable in the context of a table and because there's already
a `field_name` helper method in Rails. We're also making this parameter
optional in the table component so it isn't necessary when writing
tests.