Show link to comments for selected proposals
This link was hidden in commit 1a0f0215a in order to simplify the
information shown in selected proposals. However, we're still showing
the comments on that page, so having a link to the comments section is
useful. It might also be confusing for people to see this information in
some cases but not in other cases, for reasons they probably won't be
able to understand.

We aren't 100% this is the best option, but we're choosing the one that
keeps things simple, and that is using the same interface for every
proposal. This will also help when refactoring this code.