Remove performance optimizations to get comment flags
This is similar to what we did in commits b98244afd and 0214184b2. We
haven't checked for sure whether getting N+1 queries to get the comment
results in a significant impact for performance.

If necessary, we might reintroduce some kind of cache for comment flags
again in the future as long as we don't require setting it in different
controllers as we were doing until now.