Adapt "Hide" tests in order to include comment as factory
In the original "Hide" test, we visited the commentable's show page and
hid the comment there. Now, to follow the same pattern as the rest of
resources, we visit the comment's "show" page directly and hide it there.

For the rest of cases, we visit the resource's "index", but in this case,
the comment's "index" is the commentable's "show" page.

Regarding the expectations that were in the original test:
- `expect(page).not_to have_content("SPAM")` is replaced by
  `expect(page).not_to have_content resource_content`, since the "SPAM"
  text from commit 343025872b4 no longer makes sense in the test and
  what it was trying to verify was that the content was no longer there,
  which is already covered by `have_content resource_content`.
- `expect(page).not_to have_content("This comment has been deleted")`
  has been added in comments_spec.rb.

We remove the test from the original file and add these changes. In
next commits we will do the same for the following tests that
need adaptation, until they are all adapted and we can include :comment
as factory without issues.