Adapt "Hiding an resource's author" tests in order to include comment as factory
While adapting this test, we noticed that when blocking an author from
the show page of a comment, we encountered the same issue that was fixed
in commit "Fix redirection when blocking authors from legislation proposals".

Similarly, comments don't have a public index route. When blocking an
author from a comment's show page, we need to redirect to the commentable's
show page (e.g., the debate or proposal where the comment was made) instead
of trying to build a non-existent comments index route.

We've fixed this by updating index_path_options to detect when the referer
is a comment and redirect to the commentable's show page using
polymorphic_path. We've also adapted the tests to work correctly when
including :comment as a factory.