Adapt "Hide" test to include proposal_notification as factory
We remove the test from the original file and add these changes. In
next commits we will do the same for the following tests that
need adaptation, until they are all adapted and we can include
:proposal_notification as factory without issues.

Note that we are changing the minimum_interval validation so it only runs
when a new proposal notification is created. The original test used
"created_at: Date.current - 4.days" to avoid a situation where the
minimum_interval validation was also running during update actions,
and this produced a wrong notification count in system tests.

The main issue was that validate :minimum_interval was executed on both
create and update. When the code called update(moderated: true), Rails ran
all validations again, including this one. In the tests, this cause incorrect
notification counts when created_at is less than three days. The setting
"proposal_notification_minimum_interval_in_days" has a default value of three,
so notifications created within the last three days passed the validation.

Because of this, the test expected to see "Notifications (0)" but it found
"Notifications (1)" instead.

I think the correct solution is to run this validation only on create, using
validate :minimum_interval, on: :create. For me, this makes more sense
than forcing a created_at value older than three days inside the test.

Note that we are also changing update to update! in the hide_resource method
for proposal notifications. Since minimum_interval validation now only runs
on create, using update! ensures that any unexpected validation errors will
raise an exception.