Adapt "Ignore the proposal" test to include proposal_notification as factory
We remove the test from the original file and add these changes. In
next commits we will do the same for the following tests that
need adaptation, until they are all adapted and we can include
:proposal_notification as factory without issues.

Another option to adapt the test would be to add this method in
ProposalNotification:

```
  def ignored_flag?
    ignored?
  end
```

This would allow always using `be_ignored_flag` in the test without
needing the conditional.

However, I decided to keep the if/else in the test because
`ignored_flag?` seems to be reserved for resources that include the
`Flaggable` module (like Proposal, Debate, Comment, Budget::Investment),
which use the `ignored_flag_at` field.